Add market data handler, testers, bugfixes, logs
Implement the market data handler and event loop
Improve logging in all market data handler components
Add a market data handler tester with a dummy market event drainer
Add an exchange + market data handler tester
Bugfix: Fixed misplaced parenthesis causing logical bug
in TCP retransmit client connection.
Bugfix: Gap detector now updates class variable instead of
local variable on successful packet reception.
Bugfix: Added old bounds check on the reorder buffer.
Bugfix: Added consistent gap detection and reorder buffer mapping
regardless of starting sequence number.
Bugfix: All buffer_ accesses in the reorder bufffer are now properly
masked.
Bugfix: EncodedMessage decoded now uses network-to-host
to correctly decode message header fields.
Bugfix: Connnection::receive() now returns a span over the actual
payload size, not the entire buffer.
Bugfix: TCP retransmit server now correctly requests the right number
of bytes from Connection for a serialized retransmit request.
Added additional assertions as sanity checks in debug build.

+++ includes/utils/network-utils.hpp
@@ -34,7 +34,7 @@ bool set_multicast_interface(SocketFD socket_fd, const in_addr& interface) noexc
 bool enable_udp_port_sharing(SocketFD socket_fd) noexcept;
 JoinStatus join_multicast_group(SocketFD socket_fd, const char* mcast_group) noexcept;
 
-bool resize_recv_buffer(SocketFD socket_fd, size_t new_size_in_bytes) noexcept;
+bool resize_recv_buffer(SocketFD socket_fd, int new_size_in_bytes) noexcept;
 
 bool send_range(SocketFD socket_fd, const std::span<const std::byte> buffer, size_t& start, size_t end);
 // Intentional - use when the start index is not cared about
